pyproject.toml
[project]
name = "backend-template"
version = "0.1.0"
description = "backend template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "fastapi[standard]>=0.115.12",
    "pydantic-ai>=0.1.6",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "seqlog>=0.4.0",
    "sqlmodel>=0.0.24",
]
